newsquawk/requirements.txt
paho-mqtt
RPi.GPIO
boto3

#apt install python3-boto3
